requirements-dev.txt
# frozen requirements generated by pip-deepfreeze
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.5
distlib==0.3.7
docutils==0.20.1
exceptiongroup==1.1.3
filelock==3.13.0
flake8==6.1.0
flake8-isort==6.1.0
Flake8-pyproject==1.2.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
lxml-stubs==0.4.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
mypy==1.6.1
mypy-extensions==1.0.0
nh3==0.2.14
packaging==23.2
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
pyproject-api==1.6.1
pyproject_hooks==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
SecretStorage==3.3.3
tomli==2.0.1
tox==4.11.3
twine==4.0.2
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.6
zipp==3.17.0
